fix(clob): fix salt precision loss in JSON — use u32 range, side as integer
salt was generated as random u64 (max ~1.8e19) which exceeds JavaScript's
MAX_SAFE_INTEGER (9e15). Node.js JSON parser loses precision, causing the
server to compute a different EIP-712 hash than what we signed.

Fix: generate salt as u32 (max ~4.3e9, safely representable as float64).
Also: side reverted to integer 0/1 per official Polymarket API docs.

+++ src/clob/types.rs
@@ -129,7 +129,7 @@ impl Order {
             signature: None,
             maker: None,
             signer: None,
-            salt: U256::from(rand::random::<u64>()),
+            salt: U256::from(rand::random::<u32>()),
             nonce: super::signing::DEFAULT_CLOB_NONCE,
             expiration: 0,
             fee_rate_bps: None,
